Migrate listRules printing code to pkg/printer
