+++ pkg/printer/printer.go
@@ -1,9 +1,11 @@
 package printer
 
 import (
+	"encoding/json"
 	"fmt"
 	"os"
 	"strings"
+	"text/tabwriter"
 
 	"github.com/alecthomas/chroma/quick"
 	"github.com/mitchellh/colorstring"
@@ -175,5 +177,60 @@ func (p *Printer) PrintComparisonResult(results []rules.NamespaceChange, verbose
 
 	fmt.Println()
 	fmt.Printf("Diff Summary: %v Groups Created, %v Groups Updated, %v Groups Deleted\n", created, updated, deleted)
+	return nil
+}
+
+func (p *Printer) PrintRuleSet(rules map[string][]rwrulefmt.RuleGroup, format string) error {
+	type namespaceAndRuleGroup struct {
+		Namespace string `json:"namespace" yaml:"namespace"`
+		RuleGroup string `json:"rulegroup" yaml:"rulegroup"`
+	}
+	var items []namespaceAndRuleGroup
+
+	for ns, rulegroups := range rules {
+		for _, rg := range rulegroups {
+			items = append(items, namespaceAndRuleGroup{
+				Namespace: ns,
+				RuleGroup: rg.Name,
+			})
+		}
+	}
+
+	switch format {
+	case "json":
+		output, err := json.Marshal(items)
+		if err != nil {
+			return err
+		}
+
+		// go-text-template
+		if !p.disableColor {
+			return quick.Highlight(os.Stdout, string(output), "json", "terminal", "swapoff")
+		}
+
+		fmt.Print(string(output))
+	case "yaml":
+		output, err := yaml.Marshal(items)
+		if err != nil {
+			return err
+		}
+
+		// go-text-template
+		if !p.disableColor {
+			return quick.Highlight(os.Stdout, string(output), "yaml", "terminal", "swapoff")
+		}
+
+		fmt.Print(string(output))
+	default:
+		w := tabwriter.NewWriter(os.Stdout, 0, 0, 1, ' ', tabwriter.Debug)
+
+		fmt.Fprintln(w, "Namespace\t Rule Group")
+		for _, item := range items {
+			fmt.Fprintf(w, "%s\t %s\n", item.Namespace, item.RuleGroup)
+		}
+
+		w.Flush()
+	}
+
 	return nil
 }
